refactor: move request comment to common dto
