+++ src/common/dto/index.ts
@@ -1,1 +1,2 @@
 export * from './error.dto';
+export * from './request-comment.dto';
