+++ src/connection-request/dto/connection-request.dto.ts
@@ -15,6 +15,7 @@ import {
   IsUUID,
   ValidateNested,
 } from 'class-validator';
+import { RequestCommentDto } from 'src/common/dto';
 
 export class DatabaseInfoDto {
   @ApiProperty({
@@ -280,55 +281,6 @@ export class RequestDto {
   comments?: RequestCommentDto[];
 }
 
-export class RequestCommentDto {
-  @ApiProperty({
-    description: 'Incoming request identifier',
-    format: 'uuid',
-    example: '8b7e2f36-9217-4ea0-8d6e-b621fb6e5230',
-  })
-  @IsUUID()
-  requestId!: string;
-
-  @ApiProperty({
-    description: 'Comment identifier',
-    format: 'uuid',
-    example: '8b7e2f36-9217-4ea0-8d6e-b621fb6e5230',
-  })
-  @IsUUID()
-  @IsOptional()
-  commentId?: string;
-
-  @ApiProperty({
-    description: 'Author identifier (userId)',
-    format: 'uuid',
-    example: '8b7e2f36-9217-4ea0-8d6e-b621fb6e2230',
-  })
-  @IsUUID()
-  authorId!: string;
-
-  @ApiProperty({
-    description: 'Name of the author of the comment',
-    example: 'Owner User',
-  })
-  @IsDefined()
-  @IsString()
-  authorName!: string;
-
-  @ApiProperty({
-    type: Date,
-    description: 'The date the comment was created',
-  })
-  @IsDate()
-  @Type(() => Date)
-  createdDate: Date;
-
-  @ApiProperty({
-    description: 'Comment content',
-    example: 'This is a comment.',
-  })
-  @IsString()
-  content!: string;
-}
 export class ConnectionRequestProjectInfoDto {
   @ApiProperty({
     description: 'The name of the project',
